refactor: simplify world state exports and update consumers
